bootspec: remove duplicate `Generation_` enum
This was originally necessary in order to prevent infinite recursion,
but since `bootspec::v1::GenerationV1` derives `serde::Serialize`, we
can just turn it into a `serde_json::Value`.

This means we now only have to keep the
`bootspec::generation::Generation` enum in sync with its match arms in
the `Serialize` and `Deserialize` impls.