feat: read the full specialisation schema from bootspec
This fixes https://github.com/DeterminateSystems/bootspec/issues/147 by changing
the type of specialisations from `Generation` to `BootJson`.
This means that we capture the full recursive JSON schema for specialisations,
including extensions, while the previous type threw this information away.

This allows for instance to set specific sort keys for specialisations.
Such sort keys are allowed by the bootspec specification, but are
currently not retained when parsing a bootspec document.